Fix incorrect numeric type.

+++ include/bitcoin/explorer/utility/utility.hpp
@@ -272,7 +272,7 @@ script_type script_to_raw_data_script(const script_type& script);
  * @return                  True if successful.
  */
 bool sign_transaction(data_chunk& signature, const tx_type& transaction,
-    size_t index, const script_type& script, const ec_secret& secret,
+    uint32_t index, const script_type& script, const ec_secret& secret,
     const data_chunk& nonce, uint32_t hash_type=sighash::single);
 
 /**
